done commit

+++ Stormy-Weather-SwiftUI/Model/Weathers.swift
@@ -5,7 +5,7 @@ import SwiftUI
 struct CityWeather: Hashable, Codable, Identifiable {
     var id: Int
     var name: String
-    var temp: Double
+    var temp: Int
     var humidity: Int
     var description: String
     var isFavorite: Bool
